ScorerClassification has better handling of weird model_output + tests (#336)
Following cases of model_output yield following output now:
- len(model_output) == 0: No error, returns empty list for every score
- otherwise: require model_output to have 2 dimensions, with the second dimension being >= 1. If the second dimension is zero, an error is raised.
- if the second dimension == 1, it is assumed to be a binary classification problem with the `model_output` vector denoting
the probability of class 0. Thus we set `prob_class_1 =  1 - prob_class_0` and concat it to the `model_output` to get `model_output.shape == (n, 2)`.
For the case of the detection model setting  `model_output` to a one-hot encoded vector of all ones (i.e. `model_output = np.ones(shape=(n,1))` ), this does not change the values of the scores, they are still all 0.

Wrote unittests to test different number of dimensions (a.k.a. tensor order) and different sizes for each dimension. Also wrote tests for the binary classifier with hardcoded specific values.