Refactor projection and prediction heads (#474)
The projection and prediction heads are reformulated as concrete
versions of an abstract class `ProjectionHead`. The interface to the
model implementations stays the same except for:

- `BarlowTwins`
- `NNCLR`
- `SimSiam`

where the argument `num_mlp_layers` is removed.

To use any of the above models with a two-layer projection head, 
simply replace the three-layer projection head after initialization.