Add SimCLR implementation with dataset transforms (#1068)
Add SimCLR implementation with transforms and collate separated into two different blocks. In particular, at the moment, the collate function is only applied once a full batch of full-sized images is loaded. With this implementation, transforms are applied as soon as each image is loading. This allows storing in memory already the cropped images. This leads to faster training that scales with the cropping ratio. Adds:
- New collate function
- Transformation for SimCLR
- Fixes and deprecations
- Unit tests