Refactor DINOLoss to not use Center module
* Fix issue that DINOLoss.to(device) didn't copy the center tensor in the Center module to the correct device

The issue was that module.to() copies tensors instead of modfiying them inplace. There is a global pytorch flag to change this but I don't think we should change global state for this. I did some more checks and there is no simple way to keep the buffers in DINOLoss.center and Center synchronized. So instead I refactored the code a bit and removed Center from DINOLoss for now.